Production release ( for 23 of September 2016)

+++ docroot/modules/custom/ymca_google/src/GooglePush.php
@@ -10,6 +10,7 @@ use Drupal\Core\Entity\Query\QueryFactory;
 use Drupal\Core\Logger\LoggerChannelFactoryInterface;
 use Drupal\Core\Logger\LoggerChannelInterface;
 use Drupal\ymca_groupex_google_cache\Entity\GroupexGoogleCache;
+use Drupal\ymca_groupex_google_cache\GroupexGoogleCacheInterface;
 
 /**
  * Class GooglePush.
@@ -118,6 +119,13 @@ class GooglePush {
   protected $isProduction;
 
   /**
+   * The entity storage.
+   *
+   * @var \Drupal\Core\Entity\EntityStorageInterface
+   */
+  protected $cacheStorage;
+
+  /**
    * GooglePush constructor.
    *
    * @param GcalGroupexWrapperInterface $data_wrapper
@@ -129,7 +137,7 @@ class GooglePush {
    * @param EntityTypeManager $entity_type_manager
    *   Entity type manager.
    * @param DrupalProxy $proxy
-   *   Proxy.
+   *   The Proxy.
    * @param QueryFactory $query
    *   Query factory.
    */
@@ -143,6 +151,7 @@ class GooglePush {
     $this->query = $query;
 
     $this->isProduction = $this->configFactory->get('ymca_google.settings')->get('is_production');
+    $this->cacheStorage = $this->entityTypeManager->getStorage(GcalGroupexWrapper::ENTITY_TYPE);
 
     // Get the API client and construct the service object.
     $this->googleClient = $this->getClient();
@@ -154,6 +163,23 @@ class GooglePush {
    * Proceed all events collected by add methods.
    */
   public function proceed() {
+    $api_version = $this->dataWrapper->settings->get('api_version');
+    switch ($api_version) {
+      case 1:
+        $this->proceedLegacy();
+        break;
+
+      case 2:
+        $this->proceedCurrent();
+        break;
+
+    }
+  }
+
+  /**
+   * Legacy version of proceed().
+   */
+  protected function proceedLegacy() {
     $data = $this->dataWrapper->getProxyData();
 
     foreach ($data as $op => $entities) {
@@ -174,13 +200,24 @@ class GooglePush {
         }
 
         $gcal_id = $this->getCalendarIdByName($entity->field_gg_location->value);
+
+        // All items should be inserted in TESTING calendar in testing mode.
+        if (!$this->isProduction) {
+          $gcal_id = $this->getCalendarIdByName(self::TEST_CALENDAR_NAME);
+        }
+
         if (!$gcal_id) {
-          // Failed to get calendar ID. All errors are logged. Continue with next event.
+          // Failed to get calendar ID. Continue with next event.
           continue;
         }
 
         switch ($op) {
           case 'update':
+            // Do not update entities in testing mode.
+            if (!$this->isProduction) {
+              break;
+            }
+
             $event = $this->drupalEntityToGcalEvent($entity);
             if (!$event) {
               break;
@@ -222,7 +259,7 @@ class GooglePush {
                   [
                     '%op' => $op,
                     '%uri' => $entity->toUrl('canonical', ['absolute' => TRUE])->toString(),
-                    '%message' => $e->getMessage()
+                    '%message' => $e->getMessage(),
                   ]
                 );
                 $this->logStats($op, $processed);
@@ -241,6 +278,10 @@ class GooglePush {
             break;
 
           case 'delete':
+            // Do not delete entities in testing mode.
+            if (!$this->isProduction) {
+              break;
+            }
             try {
               $this->calEvents->delete(
                 $gcal_id,
@@ -275,7 +316,7 @@ class GooglePush {
                   [
                     '%op' => $op,
                     '%uri' => $entity->toUrl('canonical', ['absolute' => TRUE])->toString(),
-                    '%message' => $e->getMessage()
+                    '%message' => $e->getMessage(),
                   ]
                 );
                 $this->logStats($op, $processed);
@@ -330,7 +371,7 @@ class GooglePush {
                   [
                     '%op' => $op,
                     '%uri' => $entity->toUrl('canonical', ['absolute' => TRUE])->toString(),
-                    '%message' => $e->getMessage()
+                    '%message' => $e->getMessage(),
                   ]
                 );
                 $this->logStats($op, $processed);
@@ -357,6 +398,267 @@ class GooglePush {
     // Mark this step as done in the schedule.
     $this->dataWrapper->next();
 
+  }
+
+  /**
+   * Current version of proceed().
+   */
+  protected function proceedCurrent() {
+    $data = $this->dataWrapper->getProxyData();
+
+    // Insert.
+    Timer::start('insert');
+    $processed['insert'] = 0;
+    foreach ($data['insert'] as $entity) {
+      $event = $this->createEvent($entity);
+      if ($event === FALSE) {
+        // No children, skip for now.
+        continue;
+      }
+      $gcal_id = $this->getCalendarIdByName(self::TEST_CALENDAR_NAME);
+
+      try {
+        $created = $this->calEvents->insert($gcal_id, $event);
+        $entity->set('field_gg_gcal_id', $created->getId());
+        $entity->set('field_gg_need_up', 0);
+        // @todo save event start timestamp to field_gg_ts_utc (in UTC).
+        $entity->save();
+
+        $msg = 'Gcal event %gcal_id created from parent entity %parent_id.';
+        $this->logger->info(
+          $msg,
+          [
+            '%gcal_id' => $created->getId(),
+            '%parent_id' => $entity->id(),
+          ]
+        );
+      }
+      catch (\Google_Service_Exception $e) {
+        if ($e->getCode() == 403) {
+          $message = 'Google_Service_Exception [%op]: %message';
+          $this->logger->error(
+            $message,
+            [
+              '%message' => $e->getMessage(),
+              '%op' => 'insert',
+            ]
+          );
+          $this->logStats('insert', $processed);
+          if (strstr($e->getMessage(), 'Rate Limit Exceeded')) {
+            // Rate limit exceeded, retry. @todo limit number of retries.
+            return;
+          }
+        }
+        else {
+          $message = 'Google Service Exception for operation %op for Entity: %uri : %message';
+          $this->loggerFactory->get('GroupX_CM')->error(
+            $message,
+            [
+              '%op' => 'insert',
+              '%uri' => $entity->toUrl('canonical', ['absolute' => TRUE])->toString(),
+              '%message' => $e->getMessage(),
+            ]
+          );
+        }
+      }
+    }
+
+    // Update.
+    Timer::start('update');
+    $processed['update'] = 0;
+    foreach ($data['update'] as $entity) {
+      // @todo Get all children that were not pushed to Google.
+
+      $event = $this->createEvent($entity);
+      if ($event === FALSE) {
+        // No children, skip for now.
+        continue;
+      }
+      $gcal_id = $this->getCalendarIdByName(self::TEST_CALENDAR_NAME);
+      // @todo. Make sure that recurrence of Google event is the same with item recurrence.
+      // @todo. Foreach each child and get appropriate instance.
+      // @todo. Update each instance with new data and save gcal ids and timestamps to child items.
+      try {
+        $updated = $this->calEvents->update(
+          $gcal_id,
+          $entity->field_gg_gcal_id->value,
+          $event
+        );
+
+        $processed['update']++;
+
+        // Saving updated entity only when it was pushed successfully.
+        $entity->set('field_gg_google_event', serialize($updated));
+        $entity->save();
+      }
+      catch (\Google_Service_Exception $e) {
+        if ($e->getCode() == 403) {
+          $message = 'Google_Service_Exception [%op]: %message';
+          $this->logger->error(
+            $message,
+            [
+              '%message' => $e->getMessage(),
+              '%op' => 'update',
+            ]
+          );
+          $this->logStats('update', $processed);
+          if (strstr($e->getMessage(), 'Rate Limit Exceeded')) {
+            // Rate limit exceeded, retry. @todo limit number of retries.
+            return;
+          }
+        }
+        else {
+          // @todo Probably we are trying to update deleted event. Should be marked to be inserted.
+          // @todo 404 error needs to be catched.
+          $message = 'Google Service Exception for operation %op for Entity: %uri : %message';
+          $this->loggerFactory->get('GroupX_CM')->error(
+            $message,
+            [
+              '%op' => 'update',
+              '%uri' => $entity->toUrl('canonical', ['absolute' => TRUE])->toString(),
+              '%message' => $e->getMessage(),
+            ]
+          );
+          $this->logStats('update', $processed);
+        }
+
+      }
+      catch (\Exception $e) {
+        $msg = '%type : Error while updating event for entity [%id]: %msg';
+        $this->logger->error($msg, [
+          '%type' => get_class($e),
+          '%id' => $entity->id(),
+          '%msg' => $e->getMessage(),
+        ]);
+      }
+    }
+
+    // Delete.
+    foreach ($data['delete'] as $item) {
+      // @todo If item has gcal_id delete the event from Google.
+      // @todo If deletion went well delete the parent item and all children.
+    }
+  }
+
+  /**
+   * Example for getting all instances of the event.
+   *
+   * @todo Remove it.
+   */
+  private function getAllInstances() {
+    $cal_id = 'Calendar ID';
+    $event_id = 'Parent Event ID';
+    $events = $this->calService->events->instances($cal_id, $event_id);
+  }
+
+  /**
+   * Example for getting all instances of the event with opts.
+   *
+   * @todo Remove it.
+   */
+  private function getAllInstancesOpts() {
+    $cal_id = 'Calendar ID';
+    $event_id = 'Parent Event ID';
+
+    $timeMin = '2016-09-29T09:00:00.000-07:00';
+    $timeMax = '2016-09-29T14:00:00.000-07:00';
+
+    $opts = [
+      'timeMin' => $timeMin,
+      'timeMax' => $timeMax,
+    ];
+
+    $events = $this->calService->events->instances($cal_id, $event_id, $opts);
+  }
+
+  /**
+   * Example for getting instance by time.
+   *
+   * @todo Remove it.
+   */
+  private function getInstance() {
+    $cal_id = 'Calendar ID';
+
+    // Event ID is parent ID + start time of the event.
+    $event_id = 'e3c5a0d8k6mkao6ns9p2mt3vkc_20160929T170000Z';
+
+    $event = $this->calService->events->get($cal_id, $event_id);
+  }
+
+  /**
+   * Prepare event for insert.
+   *
+   * Make sure you use this function only for creating new events.
+   *
+   * @param \Drupal\ymca_groupex_google_cache\GroupexGoogleCacheInterface $entity
+   *   Parent cache entity.
+   *
+   * @return \Google_Service_Calendar_Event|bool
+   *   Prepared event for pushing.
+   */
+  protected function createEvent(GroupexGoogleCacheInterface $entity) {
+    // Find the most weighted child to use it's data.
+    $children = $this->proxy->findChildren($entity->id());
+
+    if (empty($children)) {
+      return FALSE;
+    }
+
+    $weighted = $this->cacheStorage->load($children[0]);
+
+    $event = new \Google_Service_Calendar_Event();
+
+    $event->setSummary(trim($weighted->field_gg_title->value));
+    $event->setLocation(trim($weighted->field_gg_location->value));
+    $event->setDescription($this->getDescription($weighted));
+
+    // @todo Create proper start time.
+    $start = new \Google_Service_Calendar_EventDateTime();
+    $start->setDateTime('2016-09-22T10:00:00.000-07:00');
+    $start->setTimeZone('America/Los_Angeles');
+    $event->setStart($start);
+
+    // @todo Create proper end time.
+    $end = new \Google_Service_Calendar_EventDateTime();
+    $end->setDateTime('2016-09-22T10:25:00.000-07:00');
+    $end->setTimeZone('America/Los_Angeles');
+    $event->setEnd($end);
+
+    // Available types: weekly, biweekly, NULL.
+    // @todo Check how to set biweekly events.
+    $event->setRecurrence(['RRULE:FREQ=WEEKLY']);
+
+    return $event;
+  }
+
+  /**
+   * Prepares description for the Google vvent.
+   *
+   * @param \Drupal\ymca_groupex_google_cache\GroupexGoogleCacheInterface $entity
+   *   Cache entity to get basic description.
+   *
+   * @return string
+   *   Description.
+   */
+  protected function getDescription(GroupexGoogleCacheInterface $entity) {
+    // We should prepend description with instructor.
+    $instructor = trim($entity->field_gg_instructor->value);
+
+    // Check whether we have subbed one.
+    $regex = '/<span class=\"subbed\".*><br>(.*)<\/span>/';
+    preg_match($regex, $instructor, $match);
+    if (isset($match[1])) {
+      $instructor = str_replace($match[0], ' ', $instructor);
+      $instructor .= $match[1];
+    }
+
+    // Add instructor to description.
+    $description = 'Instructor: ' . $instructor . "\n\n";
+
+    // Remove garbage.
+    $description .= strip_tags(trim(html_entity_decode($entity->field_gg_description->value)));
+
+    return $description;
   }
 
   /**
@@ -447,7 +749,7 @@ class GooglePush {
    * @param GroupexGoogleCache $entity
    *   Entity.
    *
-   * @return \Google_Service_Calendar_Event
+   * @return \Google_Service_Calendar_Event|bool
    *   Event.
    */
   private function drupalEntityToGcalEvent(GroupexGoogleCache $entity) {
@@ -555,7 +857,7 @@ class GooglePush {
         $dates[] = $date_time->setTimestamp($timestamp_item)->format(self::RRULE_DATE);
       }
 
-      // Loop over the each single week and check if the date exists in the event.
+      // Loop over the each single week and check if date exists in the event.
       $exclude = [];
       $current = $start_date_time->getTimestamp();
       while ($current <= $until_date_time->getTimestamp()) {
@@ -615,7 +917,7 @@ class GooglePush {
   /**
    * Return raw list of calendars (except primary).
    *
-   * @return array
+   * @return array|bool
    *   Array of Gcal list entries.
    */
   public function getRawCalendars() {
@@ -703,7 +1005,7 @@ class GooglePush {
         $this->logger->error($message,
           [
             '%id' => $id,
-            '%msg' => $e->getMessage()
+            '%msg' => $e->getMessage(),
           ]
         );
       }
@@ -731,7 +1033,14 @@ class GooglePush {
       return $id;
     }
     catch (\Exception $e) {
-      $this->logger->error('Failed to create calendar with name: %name', ['%name' => $name]);
+      $msg = 'Failed to create calendar with name: %name. Error: %error';
+      $this->logger->error(
+        $msg,
+        [
+          '%name' => $name,
+          '%error' => $e->getMessage(),
+        ]
+      );
       return FALSE;
     }
   }
